desafio_01/pyproject.toml
[tool.poetry]
name = "desafio-01"
version = "0.1.0"
description = ""
authors = ["Lorenzo Vagliano <lorenzo.ventura.vagliano@cern.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
django = "^5.2.1"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"
celery = {extras = ["redis"], version = "^5.5.2"}
playwright = "^1.52.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
